image_captioning_backend/caption_generator_send/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
jax==0.4.8
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pip==23.0.1
protobuf==4.22.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
